request permission for writing file

+++ mobile/src/main/java/com/michaeltroger/datarecording/sensor/PersistDataTask.java
@@ -4,20 +4,20 @@ package com.michaeltroger.datarecording.sensor;
 import android.os.AsyncTask;
 import android.os.Environment;
 import android.support.annotation.NonNull;
+import android.util.Log;
 
 import com.opencsv.CSVWriter;
 
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
 // TODO: do the actual implementation
 public class PersistDataTask extends AsyncTask<Void, Void, Void> {
 
     private static final String APP_DIRECTORY = "DataRecording";
+    private static final String TAG =PersistDataTask.class.getSimpleName();
 
     private ConcurrentLinkedQueue<float[]> valuesQueue = new ConcurrentLinkedQueue<>();
     private final FileWriter fileWriter;
@@ -27,7 +27,18 @@ public class PersistDataTask extends AsyncTask<Void, Void, Void> {
         final String fileName = "test.csv"; // TODO: dynamically choose file
         final String filePath = baseDir + File.separator + APP_DIRECTORY + File.separator + fileName;
 
-        fileWriter = new FileWriter(filePath, true);
+        final File folder = new File( Environment.getExternalStorageDirectory(), APP_DIRECTORY);
+        if (!folder.exists()) {
+            folder.mkdir();
+        }
+
+        final File file = new File(filePath);
+        if (file.exists()) {
+            fileWriter = new FileWriter(filePath, true);
+        } else {
+            fileWriter = new FileWriter(filePath);
+        }
+
     }
 
     @Override
@@ -49,6 +60,7 @@ public class PersistDataTask extends AsyncTask<Void, Void, Void> {
             for (int i = 0; i < valuesAsFloat.length; i++) {
                 valuesAsString[i] = String.valueOf(valuesAsFloat[i]);
             }
+            Log.d(TAG, "persisting data");
             csvWriter.writeNext(valuesAsString);
         }
     }
